+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Added Chisel integration for Colored Redstone Blocks (Redstone group).
 - Added option to enable/disable colored names and tooltips for blocks and items.
 - New colored property - Burnable - brown items and blocks in item form can be used as a fuel. Property can be enabled/disabled in config. In addition, burning time can be set in config (default: 200 ticks - 1 item).
+- Added polish translation.
+### Changed
+- Changed options arrangement in the configuration file. **You may need to delete the old configuration file when updating to properly generate a new one!**
 ## 1.0.1
 ### Added
 - Added translations for color names.